arc-fixed-ddave_exe: Support RLE compression
This allows the CGA and VGA graphics to be extracted and replaced, however due
to the compression it is difficult to hit data that is the perfect size.
Hopefully someone will work out where the field is storing the length of the
compressed data so it can be updated upon modification to fix this problem.