arc-fixed-ddave_exe: Move RLE padding byte handling here
This extra byte turns out to be at the end of each RLE chunk, so it's missing
from files outside the archive.  As it's a side effect of the archive's
compression, it belongs in the archive so that once files are extracted, they
are in the same format as files not found inside the archive.