Add unique ID to files to avoid recompressing unmodified files.
For compressed archives, opening and saving an archive means decompressing
every file then recompressing it again, which is very slow (> 60 s in the
"real game file" unit tests, and similarly in the web browser).  To speed this
up, we set a unique ID on each file's getContent() function, so we can detect
if this function has been changed.  If so, we compress the data as normal, but
if the data hasn't been modified, we simply copy across the original data
unchanged.

This reduces the time taken to save these archives from over a minute to under
a second.